add TeArchiveFolderView
