+++ src/widgets/TeArchiveFolderView.cpp
@@ -1,0 +1,10 @@
+#include "TeArchiveFolderView.h"
+
+TeArchiveFolderView::TeArchiveFolderView(QWidget *parent)
+	: TeFolderView(parent)
+{
+}
+
+TeArchiveFolderView::~TeArchiveFolderView()
+{
+}
