+++ src/widgets/TeFileListView.h
@@ -22,7 +22,7 @@
 
 #include <QListView>
 
-class TeFileFolderView;
+class TeFolderView;
 
 class TeFileListView : public QListView
 {
@@ -32,13 +32,13 @@ public:
 	TeFileListView(QWidget *parent = Q_NULLPTR);
 	virtual  ~TeFileListView();
 
-	virtual TeFileFolderView* folderView();
-	void setFolderView(TeFileFolderView* view);
+	virtual TeFolderView* folderView();
+	void setFolderView(TeFolderView* view);
 
 protected:
 	virtual QItemSelectionModel::SelectionFlags selectionCommand(const QModelIndex &index,	const QEvent *event = Q_NULLPTR) const;
 	virtual void keyPressEvent(QKeyEvent *event) Q_DECL_OVERRIDE;
 
 private:
-	TeFileFolderView* mp_folderView;
+	TeFolderView* mp_folderView;
 };
